+++ CHANGELOG.md
@@ -12,6 +12,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
+- Fix text
+- Fix youtube
+- Update index for Country/City
+- Update flask
+- Update bsky
+- Update text script
+
 ### Removed